Fix some test failures
Output a message that the connection failure messages are an
expected part of the test.

Increase timeouts from 30 seconds to 2 minutes, as some
cpantesters have failed on the timeouts.  So far, I'm going to
assume that the relevant systems are simply slow test machines.

One test depends on the text returned by strerror when a broken
pipe occurs.  It was failing on systems with a non-English locale.
Setting the locale to C hopefully fixes that.

Tests which need git should skip if a working git is not in PATH.