Relax warnings test
We expect an "ssh exited" and/or "broken pipe" warning here, but
it's not guaranteed to get both of them (it depends when events
arrive in the event loop).

Additionally, for the "Broken pipe" warning, the provided string
depends on the OS and locale, so attempting to test it precisely
will fail on some platforms.

Therefore, loosen the test to simply check that 1 or 2 warnings
occurred.